feat(order-service): update-convert-type-in-dto-request-order

+++ order-service/src/main/java/com/ecommerce/order_service/dto/request/OrderRequest.java
@@ -24,6 +24,6 @@ public class OrderRequest {
     LocalDate createAt;
     LocalDate updateAt;
     @Min(value = 0, message = "total price must equal or greater than 0")
-    BigDecimal totalPrice;
+    int totalPrice;
     String status;
 }
